Fixed: Move stuff back in to plugins
