+++ src/plugins/RelightThreeCanvas.js
