+++ src/plugins/RelightTorchButton.js
